requirements.txt
fastapi
httpx
pyyaml
uvicorn